+++ enjoy.py
@@ -94,7 +94,9 @@ def main():  # noqa: C901
     if not found:
         raise ValueError(f"No model found for {algo} on {env_id}, path: {model_path}")
 
-    if algo in ["dqn", "ddpg", "sac", "td3", "tqc"]:
+    off_policy_algos = ["dqn", "ddpg", "sac", "her", "td3", "tqc"]
+
+    if algo in off_policy_algos:
         args.n_envs = 1
 
     set_random_seed(args.seed)
@@ -135,7 +137,7 @@ def main():  # noqa: C901
     )
 
     kwargs = dict(seed=args.seed)
-    if algo in ["dqn", "ddpg", "sac", "her", "td3", "tqc"]:
+    if algo in off_policy_algos:
         # Dummy buffer size as we don't need memory to enjoy the trained agent
         kwargs.update(dict(buffer_size=1))
 
@@ -144,7 +146,7 @@ def main():  # noqa: C901
     obs = env.reset()
 
     # Force deterministic for DQN, DDPG, SAC and HER (that is a wrapper around)
-    deterministic = args.deterministic or algo in ["dqn", "ddpg", "sac", "her", "td3", "tqc"] and not args.stochastic
+    deterministic = args.deterministic or algo in off_policy_algos and not args.stochastic
 
     state = None
     episode_reward = 0.0
@@ -198,13 +200,13 @@ def main():  # noqa: C901
                     episode_reward, ep_len = 0.0, 0
 
     if args.verbose > 0 and len(successes) > 0:
-        print("Success rate: {:.2f}%".format(100 * np.mean(successes)))
+        print(f"Success rate: {100 * np.mean(successes):.2f}%")
 
     if args.verbose > 0 and len(episode_rewards) > 0:
-        print("Mean reward: {:.2f} +/- {:.2f}".format(np.mean(episode_rewards), np.std(episode_rewards)))
+        print(f"Mean reward: {np.mean(episode_rewards):.2f} +/- {np.std(episode_rewards):.2f}")
 
     if args.verbose > 0 and len(episode_lengths) > 0:
-        print("Mean episode length: {:.2f} +/- {:.2f}".format(np.mean(episode_lengths), np.std(episode_lengths)))
+        print(f"Mean episode length: {np.mean(episode_lengths):.2f} +/- {np.std(episode_lengths):.2f}")
 
     # Workaround for https://github.com/openai/gym/issues/893
     if not args.no_render: