+++ utils/utils.py
@@ -2,13 +2,14 @@ import argparse
 import glob
 import importlib
 import os
-from typing import Dict, Tuple
+from typing import Any, Callable, Dict, List, Optional, Tuple, Union
 
 import gym
 import yaml
 from stable_baselines3 import A2C, DDPG, DQN, HER, PPO, SAC, TD3
+from stable_baselines3.common.callbacks import BaseCallback
 from stable_baselines3.common.env_util import make_vec_env
-from stable_baselines3.common.vec_env import DummyVecEnv, SubprocVecEnv, VecFrameStack, VecNormalize
+from stable_baselines3.common.vec_env import DummyVecEnv, SubprocVecEnv, VecEnv, VecFrameStack, VecNormalize
 
 try:
     from sb3_contrib import TQC  # pytype: disable=import-error
@@ -32,7 +33,7 @@ if TQC is not None:
     ALGOS["tqc"] = TQC
 
 
-def flatten_dict_observations(env):
+def flatten_dict_observations(env: gym.Env) -> gym.Env:
     assert isinstance(env.observation_space, gym.spaces.Dict)
     try:
         return gym.wrappers.FlattenObservation(env)
@@ -41,7 +42,7 @@ def flatten_dict_observations(env):
         return gym.wrappers.FlattenDictWrapper(env, dict_keys=list(keys))
 
 
-def get_wrapper_class(hyperparams):
+def get_wrapper_class(hyperparams: Dict[str, Any]) -> Optional[Callable[[gym.Env], gym.Env]]:
     """
     Get one or more Gym environment wrapper class specified as a hyper parameter
     "env_wrapper".
@@ -55,9 +56,9 @@ def get_wrapper_class(hyperparams):
         - utils.wrappers.TimeFeatureWrapper
 
 
-    :param hyperparams: (dict)
-    :return: a subclass of gym.Wrapper (class object) you can use to
-             create another Gym env giving an original env.
+    :param hyperparams:
+    :return: maybe a callable to wrap the environment
+        with one or multiple gym.Wrapper
     """
 
     def get_module_name(wrapper_name):
@@ -98,10 +99,10 @@ def get_wrapper_class(hyperparams):
             wrapper_classes.append(wrapper_class)
             wrapper_kwargs.append(kwargs)
 
-        def wrap_env(env):
+        def wrap_env(env: gym.Env) -> gym.Env:
             """
-            :param env: (gym.Env)
-            :return: (gym.Env)
+            :param env:
+            :return:
             """
             for wrapper_class, kwargs in zip(wrapper_classes, wrapper_kwargs):
                 env = wrapper_class(env, **kwargs)
@@ -112,7 +113,7 @@ def get_wrapper_class(hyperparams):
         return None
 
 
-def get_callback_list(hyperparams):
+def get_callback_list(hyperparams: Dict[str, Any]) -> List[BaseCallback]:
     """
     Get one or more Callback class specified as a hyper-parameter
     "callback".
@@ -125,8 +126,8 @@ def get_callback_list(hyperparams):
         - utils.callbacks.PlotActionWrapper
         - stable_baselines3.common.callbacks.CheckpointCallback
 
-    :param hyperparams: (dict)
-    :return: (List[BaseCallback])
+    :param hyperparams:
+    :return:
     """
 
     def get_module_name(callback_name):
@@ -170,23 +171,33 @@ def get_callback_list(hyperparams):
 
 
 def create_test_env(
-    env_id, n_envs=1, stats_path=None, seed=0, log_dir="", should_render=True, hyperparams=None, env_kwargs=None
-):
+    env_id: str,
+    n_envs: int = 1,
+    stats_path: Optional[str] = None,
+    seed: int = 0,
+    log_dir: Optional[str] = None,
+    should_render: bool = True,
+    hyperparams: Optional[Dict[str, Any]] = None,
+    env_kwargs: Optional[Dict[str, Any]] = None,
+) -> VecEnv:
     """
     Create environment for testing a trained agent
 
-    :param env_id: (str)
-    :param n_envs: (int) number of processes
-    :param stats_path: (str) path to folder containing saved running averaged
-    :param seed: (int) Seed for random number generator
-    :param log_dir: (str) Where to log rewards
-    :param should_render: (bool) For Pybullet env, display the GUI
-    :param hyperparams: (dict) Additional hyperparams (ex: n_stack)
-    :param env_kwargs: (Dict[str, Any]) Optional keyword argument to pass to the env constructor
-    :return: (gym.Env)
+    :param env_id:
+    :param n_envs: number of processes
+    :param stats_path: path to folder containing saved running averaged
+    :param seed: Seed for random number generator
+    :param log_dir: Where to log rewards
+    :param should_render: For Pybullet env, display the GUI
+    :param hyperparams: Additional hyperparams (ex: n_stack)
+    :param env_kwargs: Optional keyword argument to pass to the env constructor
+    :return:
     """
     # Create the environment and wrap it if necessary
     env_wrapper = get_wrapper_class(hyperparams)
+
+    hyperparams = {} if hyperparams is None else hyperparams
+
     if "env_wrapper" in hyperparams.keys():
         del hyperparams["env_wrapper"]
 
@@ -214,7 +225,7 @@ def create_test_env(
     if stats_path is not None:
         if hyperparams["normalize"]:
             print("Loading running average")
-            print("with params: {}".format(hyperparams["normalize_kwargs"]))
+            print(f"with params: {hyperparams['normalize_kwargs']}")
             path_ = os.path.join(stats_path, "vecnormalize.pkl")
             if os.path.exists(path_):
                 env = VecNormalize.load(path_, env)
@@ -231,7 +242,7 @@ def create_test_env(
     return env
 
 
-def linear_schedule(initial_value):
+def linear_schedule(initial_value: Union[float, str]) -> Callable[[float], float]:
     """
     Linear learning rate schedule.
 
@@ -241,13 +252,13 @@ def linear_schedule(initial_value):
     if isinstance(initial_value, str):
         initial_value = float(initial_value)
 
-    def func(progress):
+    def func(progress_remaining: float) -> float:
         """
         Progress will decrease from 1 (beginning) to 0
-        :param progress: (float)
+        :param progress_remaining: (float)
         :return: (float)
         """
-        return progress * initial_value
+        return progress_remaining * initial_value
 
     return func
 
@@ -264,21 +275,21 @@ def get_trained_models(log_folder: str) -> Dict[str, Tuple[str, str]]:
         for env_id in os.listdir(os.path.join(log_folder, algo)):
             # Retrieve env name
             env_id = env_id.split("_")[0]
-            trained_models["{}-{}".format(algo, env_id)] = (algo, env_id)
+            trained_models[f"{algo}-{env_id}"] = (algo, env_id)
     return trained_models
 
 
-def get_latest_run_id(log_path, env_id):
+def get_latest_run_id(log_path: str, env_id: str) -> int:
     """
     Returns the latest run number for the given log name and log path,
     by finding the greatest number in the directories.
 
-    :param log_path: (str) path to log folder
-    :param env_id: (str)
-    :return: (int) latest run number
+    :param log_path: path to log folder
+    :param env_id:
+    :return: latest run number
     """
     max_run_id = 0
-    for path in glob.glob(log_path + "/{}_[0-9]*".format(env_id)):
+    for path in glob.glob(log_path + f"/{env_id}_[0-9]*"):
         file_name = path.split("/")[-1]
         ext = file_name.split("_")[-1]
         if env_id == "_".join(file_name.split("_")[:-1]) and ext.isdigit() and int(ext) > max_run_id:
@@ -286,12 +297,12 @@ def get_latest_run_id(log_path, env_id):
     return max_run_id
 
 
-def get_saved_hyperparams(stats_path, norm_reward=False, test_mode=False):
+def get_saved_hyperparams(stats_path: str, norm_reward: bool = False, test_mode: bool = False) -> Tuple[Dict[str, Any], str]:
     """
-    :param stats_path: (str)
-    :param norm_reward: (bool)
-    :param test_mode: (bool)
-    :return: (dict, str)
+    :param stats_path:
+    :param norm_reward:
+    :param test_mode:
+    :return:
     """
     hyperparams = {}
     if not os.path.isdir(stats_path):