+++ CHANGELOG.md
@@ -12,6 +12,8 @@
 ### Other
 - Refactored train script, now uses a `ExperimentManager` class
 - Replaced `make_env` with SB3 built-in `make_vec_env`
+- Add more type hints (`utils/utils.py` done)
+- Use f-strings when possible
 
 ## Pre-Release 0.10.0 (2020-10-28)